Refactor the Atoms and Atom definitions
